infrastructure/requirements.txt
attrs==22.1.0
aws-cdk-lib==2.44.0
cattrs==22.1.0
constructs==10.1.116
dynaconf==3.1.11
exceptiongroup==1.0.0rc9
jsii==1.69.0
publication==0.0.3
python-dateutil==2.8.2
six==1.16.0
typeguard==2.13.3
typing-extensions==4.3.0
